requirements.txt
tkinter
numpy
matplotlib
astropy
spiceypy
skyfield
cupy-cuda12x